requirements.txt
chromedriver-binary==86.0.4240.22.0
flake8==3.8.4
isort==5.6.4
mccabe==0.6.1
mimesis==4.1.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
selenium==3.141.0
slackweb==1.0.5
urllib3==1.25.11
